backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
python-docx
google-generativeai
shutil
